move to instance_configs
